move buffer-actions from buffers
